pyproject.toml
[tool.poetry]
name = "consul-docker-autosync"
version = "0.1.0"
description = ""
authors = ["Dani Hodovic <dani.hodovic@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
docker = "^5.0.0"
python-consul2 = "^0.1.5"
click = "^7.1.2"
loguru = "^0.5.3"
Flask = "^1.1.2"
waitress = "^2.0.0"

[tool.poetry.dev-dependencies]
pytest = "^5.2"
black = "^20.8b1"
pylint = "^2.7.4"
mypy = "^0.812"
pudb = "^2020.1"
ptipython = "^1.0.1"
requests = "^2.25.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
cli = "consul_docker_autosync.cli:run"
